move src/text to src/internal/text
