+++ src/internal/text/TextCaption.tsx
