+++ src/internal/text/TextLegal.tsx
