+++ src/internal/text/index.ts
@@ -1,0 +1,6 @@
+export { TextBody } from './TextBody';
+export { TextCaption } from './TextCaption';
+export { TextHeadline } from './TextHeadline';
+export { TextLabel1 } from './TextLabel1';
+export { TextLabel2 } from './TextLabel2';
+export { TextLegal } from './TextLegal';
